resolved conflict

+++ src/RadMarshakCGS/test_radiation_marshak_cgs.cpp
@@ -311,7 +311,7 @@ auto problem_main() -> int
 		const double t = sim.tNew_[0];
 		const double xmax = c_light_cgs_ * t;
 		amrex::Print() << "diffusion length = " << xmax << std::endl;
-		for (int i = 0; i < xs_exact.size(); ++i) {
+		for (size_t i = 0; i < xs_exact.size(); ++i) {
 			if (xs_exact[i] < xmax) {
 				err_norm += std::abs(Trad_interp[i] - Trad_exact[i]);
 				sol_norm += std::abs(Trad_exact[i]);
